+++ src/components/Table.js
@@ -1,6 +1,7 @@
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
+import { fetchData, filterPlanet } from '../action/index';
 
 export class Table extends Component {
 constructor(props){
@@ -9,6 +10,17 @@ constructor(props){
   this.filtraData = this.filtraData.bind(this);
   this.filterDataByName = this.filterDataByName.bind(this);
   this.renderTableBody = this.renderTableBody.bind(this);
+  this.fetchUrl = this.fetchUrl.bind(this);
+}
+
+componentDidMount() {
+  this.fetchUrl();
+}
+
+fetchUrl() {
+  const { request } = this.props;
+  request();
+  console.log(this.props.value)
 }
 
 retornaSign(comparison) {
@@ -25,7 +37,6 @@ retornaSign(comparison) {
   }
   return null;
 }
-
 
 filtraData(comparisonSignal, results, column, value, name) {
   if (comparisonSignal === 0) {
@@ -54,11 +65,31 @@ filterDataByName(filtername, results) {
   renderTableBody(filtername, filterNumb) {
     const { value: { data } } = this.props;
     const { results } = data;
-     filterNumb.filterByNumericValues.map((element) => {
+     return filterNumb.filterByNumericValues.map((element) => {
       const signal = this.retornaSign(element.comparison);
-      this.filteredPlanets =  this.filtraData(signal, results, element.column, element.value, filtername);
-    });  
-    this.filteredPlanets.map((element) => (
+      return  this.filtraData(signal, results, element.column, element.value, filtername);
+    });
+  }
+
+  render() {
+    const { value: { data } } = this.props;
+    const { value: { filters: { filterByName: { name } } } } = this.props;
+    const { value: { filters } } = this.props;
+    const { results } = data;
+    if(this.filteredPlanets === undefined ) {
+    this.filteredPlanets = results;
+    console.log(this.filteredPlanets)}
+    const headers = ['name', 'rotation_period', 'orbital_period', 'diameter', 'climate', 'gravity', 'terrain', 'surface_water', 'population', 'films', 'created', 'edited', 'url'];
+    return (
+      <div>
+         <table>
+                <thead>
+                  <tr>
+                    {headers.map((element) => <th key={element}>{element}</th>)}
+                  </tr>
+                </thead>
+                <tbody>
+                  {results.map((element) => (
           <tr key={element.name}>
             <td>{element.name}</td>
             <td>{element.rotation_period}</td>
@@ -73,55 +104,27 @@ filterDataByName(filtername, results) {
             <td>{element.created}</td>
             <td>{element.edited}</td>
             <td>{element.url}</td>
-          </tr>
-        )) 
-        console.log(this.filteredPlanets)    
-  }
-
-  render() {
-    const { value: { data } } = this.props;
-    const { value: { filters: { filterByName: { name } } } } = this.props;
-    const { value: { filters } } = this.props;
-    const { results } = data;
-    if(this.filteredPlanets === undefined ) {
-    this.filteredPlanets = results;
-    console.log(this.filteredPlanets)}
-    const headers = ['name', 'rotation_period', 'orbital_period', 'diameter', 'climate', 'gravity', 'terrain', 'surface_water', 'population', 'films', 'created', 'edited', 'url'];
-    return (
-      <div>
-        {results.length === 1
-          ? (
-            <h1>
-              {results[0]}
-            </h1>
-          ) : (
-            <div>
-              <p> Table </p>
-              <table>
-                <thead>
-                  <tr>
-                    {headers.map((element) => <th key={element}>{element}</th>)}
-                  </tr>
-                </thead>
-                <tbody>
-                  {this.renderTableBody(name, filters)}
-                </tbody>
+         </tr>))}
+             </tbody>
               </table>
             </div>
-          )}
-      </div>
-    );
-  }
-}
+ )}}
 
 const mapStateToProps = (state) => ({ value: state });
 
+const mapDispatchToProps = (dispatch) => ({
+  request: (e) => dispatch(fetchData(e)),
+  filter: (e) => dispatch(filterPlanet(e)),
+});
+
 Table.propTypes = {
   value: PropTypes.instanceOf(Object),
+  request: PropTypes.func,
 };
 
 Table.defaultProps = {
   value: {},
+  request: PropTypes.func,
 };
 
-export default connect(mapStateToProps)(Table);
+export default connect(mapStateToProps,mapDispatchToProps)(Table);
